+++ doc/Changelog.md
@@ -3,7 +3,8 @@ Changelog {#changelog}
 
 # git master {#master}
 
-* [#77](https://github.com/BlueBrain/Fivox/pull/77)
+* [#77](https://github.com/BlueBrain/Fivox/pull/77),
+  [#78](https://github.com/BlueBrain/Fivox/pull/78)
   Enable CUDA implementation of the lfp computation.
 
 # Release 0.6 (09-12-2016) {#Release06}